defaulting to float percision in procesing functions

+++ examples/Gain/GainProcessor.h
@@ -13,7 +13,7 @@ public:
     ~GainProcessor(){};
 
     // Process method to handle audio processing
-    void process(double** inputs, double** outputs, std::size_t blockSize) override;
+    void process(float** inputs, float** outputs, std::size_t blockSize) override;
 
     // Set a parameter value
     void setParameter(int paramId, double value) override;
